requirements.txt
requests~=2.26.0
setuptools~=57.1.0
websocket-client~=1.1.0